YouTube URL tagger: ignore Shorts
YouTube Shorts don't have clickable links, so we don't need to use UTM/URL parameters. This update disables the URL tagger button for Shorts. (It is paired with an update to the Google Sheet updater that stops Shorts being marked as having incorrect parameters if they don't have UTM tags.)